+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Exteded library refactoring - #161
 - Removed "File->Import Project" menu. Instead, project are imported now via "File->Open Project". - #202
 - New Splashscreen
+- Restructuring of batch application including the new interface for modules to add batch functions - #265
 
 ### Added
 - Added dynamic interface to share arbitrary functions with other modules. These functions need to be defined in the method GtModuleInterface::sharedFunctions - #207, #47